Document queryParams + rename example module
This file does more than transform, and `examples` seems is a bit more consistent with e.g. `trace.js` and `route.js`.
