+++ core/base-service/examples.js
